- add loopring class - update request.js & apis.ks - update validator
