+++ 2.0/common/apis.js
@@ -1,19 +1,18 @@
-import request from '../common/request'
-import validator from '../common/validator'
+import request from './request'
+import validator from './validator'
 // import {ADDRESS,RPC_TAG,HEX} from './types'
 
 let headers = {
     'Content-Type': 'application/json'
-};
-const HOST = 'https://relay1.loopring.io/rpc'
+}
 
 export async function getTransactionCount(add,tag){
   validator.validate({value:add,type:'ADDRESS'})
-
+  
   let body = {}
   body.method = 'eth_getTransactionCount'
   body.params = [add,tag]
-  return request(`${HOST}`,{
+  return request({
     method:'post',
     headers,
     body,
@@ -22,11 +21,11 @@ export async function getTransactionCount(add,tag){
 
 export async function sendRawTransaction(signedTx){
     validator.validate({value:signedTx,type:'HEX',})
-    
+
     let body = {}
     body.method = 'eth_sendRawTransaction'
     body.params = [signedTx]
-    return request(`${HOST}`,{
+    return request({
       method:'post',
       headers,
       body,
