+++ 2.0/common/httpprovider.js
@@ -1,0 +1,9 @@
+
+export default class HttpProvider {
+
+  constructor(host) { 
+    this.host = host
+
+  }
+}
+
