+++ 2.0/index.js
@@ -1,7 +1,19 @@
-require('babel-polyfill')
 import Transaction from './transaction';
+import HttpProvider from './common/httpprovider';
+import validator from './common/validator';
 
-export default {
-  Transaction
+export default class Loopring {
+
+  constructor(host) {
+  	validator.validate({type:'URL',value:host})
+  	// browser env
+    if(typeof window !== 'undefined'){
+    	window.LOOPRING_PROVIDER_HOST = host // used by apis.js
+    }
+    // node env
+    if(typeof global !== 'undefined'){
+    	global.LOOPRING_PROVIDER_HOST = host // used by apis.js
+    }
+  }
 }
