+++ 2.0/index.web3.js
@@ -1,0 +1,18 @@
+import Transaction from './transaction';
+import HttpProvider from './common/httpprovider';
+
+export default class Loopring {
+
+  constructor(provider) { 
+    this.currentProvider = provider
+    // this.transaction = new Transaction(this)
+    this.providers = {
+        HttpProvider: HttpProvider,
+    }
+  }
+  setProvider(provider) {
+  	this.currentProvider = provider
+  }
+  
+}
+
